amides/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
certifi==2022.12.7
    # via elastic-transport
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
elastic-transport==8.4.0
    # via elasticsearch
elasticsearch==8.6.2
    # via -r requirements.in
fonttools==4.39.0
    # via matplotlib
joblib==1.2.0
    # via
    #   -r requirements.in
    #   scikit-learn
kiwisolver==1.4.4
    # via matplotlib
luqum==0.12.1
    # via -r requirements.in
matplotlib==3.7.1
    # via
    #   -r requirements.in
    #   seaborn
ndjson==0.3.1
    # via -r requirements.in
numpy==1.24.2
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
packaging==23.0
    # via matplotlib
pandas==1.5.3
    # via
    #   -r requirements.in
    #   seaborn
pillow==9.4.0
    # via matplotlib
ply==3.11
    # via luqum
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.7.1
    # via pandas
pyyaml==6.0.1
    # via -r requirements.in
scikit-learn==1.2.2
    # via -r requirements.in
scipy==1.10.1
    # via
    #   -r requirements.in
    #   scikit-learn
seaborn==0.12.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
threadpoolctl==3.1.0
    # via scikit-learn
urllib3==1.26.15
    # via elastic-transport
